Refacator existing temaplte test and add new test case of connectors Ids validation
