+++ .azure-pipelines/detectionTemplateSchemaValidation.yaml
@@ -1,5 +1,5 @@
 jobs:
-- job: "DetectionTemplateStructureValidation"
+- job: "DetectionTemplateSchemaValidation"
   pool:
     vmImage: 'ubuntu-18.04'
   variables:
@@ -16,4 +16,4 @@ jobs:
         command: 'test'
         arguments: '--configuration $(buildConfiguration)'
         publishTestResults: true
-        projects: '**/DetectionTemplateStructureValidation.Tests.csproj'
+        projects: '**/DetectionTemplateSchemaValidation.Tests.csproj'
