+++ .script/tests/detectionTemplateSchemaValidation/DetectionTemplateSchemaValidation.Tests.csproj
