+++ .script/tests/detectionTemplateSchemaValidation/DetectionTemplateSchemaValidation.Tests.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30621.155
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DetectionTemplateStructureValidation.Tests", "DetectionTemplateStructureValidation.Tests.csproj", "{5B7A7358-FA15-45A7-B780-50180855845C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DetectionTemplateSchemaValidation.Tests", "DetectionTemplateSchemaValidation.Tests.csproj", "{5B7A7358-FA15-45A7-B780-50180855845C}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{C485F47B-B0C7-47C0-821A-00F7FDD5896A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
