+++ .script/tests/detectionTemplateSchemaValidation/DetectionsYamlFilesTestData.cs
