+++ .script/tests/detectionTemplateSchemaValidation/Models/AlertTriggerOperator.cs
