+++ .script/tests/detectionTemplateSchemaValidation/Models/AnalyticsTemplateInternalModelBase.cs
