+++ .script/tests/detectionTemplateSchemaValidation/Models/AttackTactic.cs
