+++ .script/tests/detectionTemplateSchemaValidation/Models/DataConnectorInternalModel.cs
