+++ .script/tests/detectionTemplateSchemaValidation/Models/EntityMapping.cs
