+++ .script/tests/detectionTemplateSchemaValidation/Models/EntityType.cs
