+++ .script/tests/detectionTemplateSchemaValidation/Models/FieldMapping.cs
