+++ .script/tests/detectionTemplateSchemaValidation/Models/FrequencyLimitationForLongPeriodQuery.cs
