+++ .script/tests/detectionTemplateSchemaValidation/Models/JsonConverterUtils.cs
