+++ .script/tests/detectionTemplateSchemaValidation/Models/NewEntityMappingsAttribute.cs
