+++ .script/tests/detectionTemplateSchemaValidation/Models/PeriodGreaterThanOrEqualFrequencyAttribute.cs
