+++ .script/tests/detectionTemplateSchemaValidation/Models/RangeTimeSpanIsoFormatAttribute.cs
