+++ .script/tests/detectionTemplateSchemaValidation/Models/ScheduledTemplateInternalModel.cs
