+++ .script/tests/detectionTemplateSchemaValidation/Models/ScheduledTemplateTimeSpanConverter.cs
