+++ .script/tests/detectionTemplateSchemaValidation/Models/ScheduledTemplateTriggerOperatorConverter.cs
