+++ .script/tests/detectionTemplateSchemaValidation/Models/ValidEntityMappingsAttribute.cs
