+++ .script/tests/detectionTemplateSchemaValidation/SkipValidationsTemplates.json
