+++ .script/tests/detectionTemplateSchemaValidation/TheoryData.cs
