+++ .script/tests/detectionTemplateSchemaValidation/ValidConnectorsIds.json
@@ -1,0 +1,106 @@
+﻿[
+  "Agari",
+  "AIVectraDetect",
+  "AkamaiSecurityEvents",
+  "AlcideKAudit",
+  "AlsidForAD",
+  "ApacheHTTPServer",
+  "ArubaClearPass",
+  "AWS",
+  "AzureActiveDirectory",
+  "AzureActiveDirectoryIdentityProtection",
+  "AzureActivity",
+  "AzureAdvancedThreatProtection",
+  "AzureFirewall",
+  "AzureInformationProtection",
+  "AzureMonitor(IIS)",
+  "AzureMonitor(VMInsights)",
+  "AzureMonitor(WireData)",
+  "AzureSecurityCenter",
+  "AzureSql",
+  "Barracuda",
+  "BarracudaCloudFirewall",
+  "BetterMTD",
+  "BeyondSecuritybeSECURE",
+  "BlackberryCylancePROTECT",
+  "BroadcomSymantecDLP",
+  "CEF",
+  "CheckPoint",
+  "CiscoASA",
+  "CiscoFirepowerEStreamer",
+  "CiscoISE",
+  "CiscoMeraki",
+  "CiscoUCS",
+  "CiscoUmbrellaDataConnector",
+  "Citrix",
+  "CitrixWAF",
+  "CyberArk",
+  "Darktrace",
+  "DDOS",
+  "DNS",
+  "Dynamics365",
+  "ESETEnterpriseInspector",
+  "EsetSMC",
+  "ExtraHopNetworks",
+  "F5",
+  "F5BigIp",
+  "ForcepointCasb",
+  "ForcepointDlp",
+  "ForcepointNgfw",
+  "ForgeRock",
+  "Fortinet",
+  "GoogleWorkspaceReportsAPI",
+  "illusiveAttackManagementSystem",
+  "ImpervaWAFGateway",
+  "ImportedConnector",
+  "InfobloxNIOS",
+  "IoT",
+  "JuniperSRX",
+  "MicrosoftCloudAppSecurity",
+  "MicrosoftDefenderAdvancedThreatProtection",
+  "MicrosoftThreatIntelligence",
+  "MicrosoftThreatProtection",
+  "Netskope",
+  "NXLogDnsLogs",
+  "NXLogLinuxAudit",
+  "Office365",
+  "OfficeATP",
+  "OktaSSO",
+  "OnapsisPlatform",
+  "OneIdentity",
+  "OrcaSecurityAlerts",
+  "PaloAltoNetworks",
+  "Perimeter81ActivityLogs",
+  "ProofpointPOD",
+  "ProofpointTAP",
+  "PulseConnectSecure",
+  "QualysKB",
+  "QualysVulnerabilityManagement",
+  "SalesforceServiceCloud",
+  "SecurityEvents",
+  "SonicWallFirewall",
+  "SophosCloudOptix",
+  "SophosXGFirewall",
+  "SquadraTechnologiesSecRmm",
+  "SquidProxy",
+  "Symantec",
+  "SymantecEndpointProtection",
+  "SymantecProxySG",
+  "SymantecVIP",
+  "Syslog",
+  "ThreatIntelligence",
+  "ThreatIntelligenceTaxii",
+  "ThycoticSecretServer_CEF",
+  "TrendMicro",
+  "TrendMicroTippingPoint",
+  "TrendMicroXDR",
+  "VMwareCarbonBlack",
+  "VMwareESXi",
+  "WAF",
+  "WatchguardFirebox",
+  "WindowsEventForwarding",
+  "WindowsFirewall",
+  "WireX_Systems_NFP",
+  "ZimperiumMtdAlerts",
+  "Zscaler"
+]
