+++ .script/tests/detectionTemplateSchemaValidation/nuget.config
