+++ azure-pipelines.yml
@@ -10,7 +10,7 @@ trigger:
 
 jobs:
 - template: .azure-pipelines/detectionsValidations.yaml
-- template: .azure-pipelines/detectionTemplateStructureValidation.yaml
+- template: .azure-pipelines/detectionTemplateSchemaValidation.yaml
 - template: .azure-pipelines/kqlValidations.yaml
 - template: .azure-pipelines/workbooksValidations.yaml
 - template: .azure-pipelines/yamlFileValidator.yaml
